IGI/LR5/medcenter/requirements.txt
Django>=5.2.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-allauth>=0.57.0
django-admin-rangefilter>=0.11.0
Pillow>=10.0.0
requests>=2.31.0